create nav graph

+++ app/src/main/java/com/example/fse_project/data/local/database/AppDao.kt
@@ -29,7 +29,7 @@ interface AppDao {
     suspend fun deleteUser(userId: Int)
 
     @Query("SELECT * FROM users WHERE id = :userId")
-    suspend fun getUserById(userId: Int): UserEntity
+    suspend fun getUserById(userId: Long): UserEntity
 
     @Query("SELECT * FROM users")
     fun getUsers(): Flow<List<UserEntity>>
