+++ app/src/main/java/com/example/fse_project/data/repository/UserRepositoryImpl.kt
@@ -34,7 +34,7 @@ class UserRepositoryImpl @Inject constructor(
         dao.deleteUser(userId)
     }
 
-    override suspend fun getUserProfile(userId: Int): User {
+    override suspend fun getUserProfile(userId: Long): User {
         return dao.getUserById(userId).toDomain()
     }
